Export inputOptions and getType functions to inputOptions file
